+++ pkg/controller/controller.go
@@ -5,6 +5,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/weaveworks/flagger/pkg/canary"
 	"github.com/weaveworks/flagger/pkg/metrics"
 	"github.com/weaveworks/flagger/pkg/router"
 
@@ -43,7 +44,7 @@ type Controller struct {
 	logger        *zap.SugaredLogger
 	canaries      *sync.Map
 	jobs          map[string]CanaryJob
-	deployer      CanaryDeployer
+	deployer      canary.Deployer
 	observer      metrics.Observer
 	recorder      metrics.Recorder
 	notifier      *notifier.Slack
@@ -63,6 +64,7 @@ func NewController(
 	routerFactory *router.Factory,
 	meshProvider string,
 	version string,
+	labels []string,
 ) *Controller {
 	logger.Debug("Creating event broadcaster")
 	flaggerscheme.AddToScheme(scheme.Scheme)
@@ -74,14 +76,15 @@ func NewController(
 	eventRecorder := eventBroadcaster.NewRecorder(
 		scheme.Scheme, corev1.EventSource{Component: controllerAgentName})
 
-	deployer := CanaryDeployer{
-		logger:        logger,
-		kubeClient:    kubeClient,
-		flaggerClient: flaggerClient,
-		configTracker: ConfigTracker{
-			logger:        logger,
-			kubeClient:    kubeClient,
-			flaggerClient: flaggerClient,
+	deployer := canary.Deployer{
+		Logger:        logger,
+		KubeClient:    kubeClient,
+		FlaggerClient: flaggerClient,
+		Labels:        labels,
+		ConfigTracker: canary.ConfigTracker{
+			Logger:        logger,
+			KubeClient:    kubeClient,
+			FlaggerClient: flaggerClient,
 		},
 	}