+++ pkg/router/factory.go
@@ -31,11 +31,12 @@ func NewFactory(kubeConfig *restclient.Config, kubeClient kubernetes.Interface,
 }
 
 // KubernetesRouter returns a ClusterIP service router
-func (factory *Factory) KubernetesRouter() *KubernetesRouter {
+func (factory *Factory) KubernetesRouter(label string) *KubernetesRouter {
 	return &KubernetesRouter{
 		logger:        factory.logger,
 		flaggerClient: factory.flaggerClient,
 		kubeClient:    factory.kubeClient,
+		label:         label,
 	}
 }